+++ remix/app.env.d.ts
@@ -1,0 +1,13 @@
+/// <reference types="vite/client" />
+
+declare global {
+  interface Window {
+    env?: Record<string, string | undefined>;
+    envFromCookie?: Record<string, string | undefined>;
+    process?: {
+      env?: Record<string, string | undefined>;
+    };
+  }
+}
+
+export {};