+++ remix/app/api/utils/mongodb/mongodb.ts
@@ -1,4 +1,3 @@
 export const getMongoDb = async () => {
-  // @ts-expect-error
   return await import('mongodb').then((mod) => mod);
 };