+++ remix/app/routes/api/v1/vercel/deployments/_deployments.tsx
@@ -1,4 +1,4 @@
-import { json } from '@vercel/remix';
+import { json } from '~/api/http';
 
 import {
   getVercelDeploymentsOverview,